scout-api/requirements-simple.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-chess==1.999
numpy==1.26.3
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
